Make APIGateway Integration call Lambda asynchronously
This is necessary because the Slack API will keep retrying messages
if they're not responded to in a timely manner - which, if your Lambda
takes a while, would lead to duplicate messages.

Also remove some unnecessary Metadata that was leftover from
the Cloudformation UI.

Also lowercase the bucket-name, since I was getting errors on uppercase -
I guess the spec changed recently?